+++ src/Fortuna.jl
@@ -15,7 +15,7 @@ using  DocStringExtensions
 # --------------------------------------------------
 # REEXPORT PACKAGES
 # --------------------------------------------------
-using Reexport
+import Reexport: @reexport
 @reexport import Distributions: rand, pdf # Extended functions
 @reexport import Distributions: mean, std, cor, params # Useful functions
 @reexport import LinearAlgebra: I
@@ -75,7 +75,8 @@ include("Random Variables/SampleRandomVariables.jl")
 include("Reliability Problems/ReliabilityProblems.jl")
 include("InverseReliabilityProblems.jl")
 include("SensitivityProblems.jl")
-include("Utilities/MakieRecipes.jl")
+# include("Utilities/MakieRecipes.jl") 
+# Temporarily removed to avoid dependency on Makie, need to create an extension
 export AbstractSamplingTechnique
 export ITS, LHS
 export AbstractTransformation